Add kwargs to the date and time classes
This allows for less code to be written by the users of this library. this way users can use the keyword arguments as expected on a date, time, datetime or timedelta object, directly into confkit data types.
Without being required to pass python's objects.
Unpack also helps with type checking.